create marshal, constants and util package refactor UnMarshalRaw
