+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/surrealdb/surrealdb.go
 
-go 1.18
+go 1.20
 
 require (
 	github.com/gorilla/websocket v1.5.0
