+++ pkg/util/util.go
@@ -1,0 +1,7 @@
+package util
+
+import "reflect"
+
+func IsSlice(value interface{}) bool {
+	return reflect.TypeOf(value).Kind() == reflect.Slice
+}
